+++ components/ReferenceManager/references/reference_uploader/ReferenceManualUploadDrawer.tsx
@@ -25,11 +25,11 @@ import Box from "@mui/material/Box";
 import CloseOutlinedIcon from "@mui/icons-material/CloseOutlined";
 import Drawer from "@mui/material/Drawer";
 import PrimaryButton from "../../form/PrimaryButton";
-import ReferenceDoiSearchInput from "./ReferenceDoiSearchInput";
+import ReferenceDoiSearchInput from "../../form/ReferenceDoiSearchInput";
 import ReferenceItemFieldCreatorTagInput from "../../form/ReferenceItemFieldCreatorTagInput";
 import ReferenceItemFieldInput from "../../form/ReferenceItemFieldInput";
 import ReferenceItemFieldSelect from "../../form/ReferenceItemFieldSelect";
-import ReferenceUploadAttachments from "./ReferenceUploadAttachments";
+import ReferenceUploadAttachments from "../../form/ReferenceUploadAttachments";
 import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
 import { useOrgs } from "~/components/contexts/OrganizationContext";
